+++ src/Components/PelcroModalController/PelcroModalController.service.js
@@ -8,6 +8,9 @@ import {
 } from "../../utils/utils";
 import { init as initContentEntitlement } from "../common/contentEntitlement";
 import { loadStripe } from "@stripe/stripe-js/pure";
+import { notify } from "../../SubComponents/Notification";
+import { getErrorMessages } from "../common/Helpers";
+import i18n from "../../i18n";
 
 /**
  * @typedef {Object} OptionsType
@@ -317,6 +320,10 @@ export const initViewFromURL = () => {
         return initCartFromUrl();
       }
 
+      if (view === "email-verify") {
+        return verifyEmailTokenFromUrl();
+      }
+
       switchView(view);
     });
   }
@@ -475,4 +482,37 @@ export const initCartFromUrl = () => {
 
     if (hasAddedAnItemSuccessfully) switchView("cart");
   });
+};
+
+const verifyEmailTokenFromUrl = () => {
+  const { whenSiteReady } = usePelcro.getStore();
+
+  const translations = i18n.t("verifyEmail:messages", {
+    returnObjects: true
+  });
+
+  const emailToken = window.Pelcro.helpers.getURLParameter("token");
+
+  const isEmailVerificationEnabled =
+    window.Pelcro.site.read()?.email_verify_enabled ?? false;
+
+  if (!emailToken || !isEmailVerificationEnabled) return;
+
+  whenSiteReady(
+    () => {
+      window.Pelcro.user.verifyEmailToken(
+        {
+          token: emailToken
+        },
+        (err, res) => {
+          if (err) {
+            return notify.error(getErrorMessages(err));
+          }
+
+          return notify.success(translations.success);
+        }
+      );
+    },
+    { once: true }
+  );
 };